requirements.txt
streamlit
pandas
numpy
plotly
google-api-python-client
google-auth-httplib2
google-auth-oauthlib